backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-socketio==5.10.0
python-multipart==0.0.6
motor==3.3.2
pymongo==4.6.0
openai==1.3.8
librosa==0.10.1
pydub==0.25.1
numpy==1.24.3
scipy==1.11.4
python-dotenv==1.0.0
aiofiles==23.2.1
websockets==12.0
pydantic==2.5.0
httpx==0.25.2
